+++ GrowableGiantCrops/Framework/ModConfig.cs
@@ -1,0 +1,14 @@
+﻿using AtraShared.Integrations.GMCMAttributes;
+
+using Microsoft.Xna.Framework;
+
+namespace GrowableGiantCrops.Framework;
+public sealed class ModConfig
+{
+    [GMCMDefaultIgnore]
+    public Vector2 ShopLocation { get; set; } = new(1, 7);
+
+    public bool ShouldNPCsTrampleGiantCrops { get; set; } = true;
+
+    public bool RelaxedPlacement { get; set; } = false;
+}