+++ PrismaticSlime/docs/Changelog.md
@@ -7,7 +7,10 @@ TODO:
 * Fix up buff indicator for the prismatic jelly toast.
 * Figure out how to get the player to glow different colors.
 * Figure out dye pots and make them free if given a prismatic jelly. Add Emily event.
-* Make crops grow seven days? -->
+* Add it as a method to make junimos change colors?
+* Add it as a new tailoring???
+* Make crops grow seven days?....
+* Qi: exchange for gems.-->
 
 ### Version 0.1.0